remove server folder, move config  and intl files to common
