+++ gulp/messages-export.js
@@ -1,13 +1,12 @@
 import fs from 'fs';
 import gulp from 'gulp';
-import loadMessages from '../src/server/intl/loadMessages';
+import loadMessages from '../src/common/intl/loadMessages';
 import { diff, messagesToCode } from './support/messages';
 
 gulp.task('messages-export', ['messages-extract'], () => {
   const messages = loadMessages({ includeDefault: true });
   const defaultMessages = messages._default; // eslint-disable-line no-underscore-dangle
-  const defaultMessagesKeys = Object
-    .keys(defaultMessages);
+  const defaultMessagesKeys = Object.keys(defaultMessages);
 
   Object.keys(messages)
     .filter(locale => locale.charAt(0) !== '_')
@@ -19,8 +18,8 @@ gulp.task('messages-export', ['messages-extract'], () => {
 
       const mergedMessages = [];
       defaultMessagesKeys.forEach(messageKey => {
-        const message = (missingMessagesKeys.indexOf(messageKey) > -1)
-          ? (`MISSING: ${defaultMessages[messageKey]}`)
+        const message = missingMessagesKeys.indexOf(messageKey) > -1
+          ? `MISSING: ${defaultMessages[messageKey]}`
           : localeMessages[messageKey];
         mergedMessages.push({
           defaultMessage: message,
