+++ package.json
@@ -3,13 +3,13 @@
   "private": true,
   "scripts": {
     "build": "next build src",
-    "dev": "node src/server/server.js",
+    "dev": "node src/server.js",
     "flow": "flow",
     "install": "gulp native-fix",
     "jest": "jest",
     "jest:watch": "jest --watch",
     "postinstall": "gulp native",
-    "start": "NODE_ENV=production node src/server/server.js",
+    "start": "NODE_ENV=production node src/server.js",
     "storybook-web": "start-storybook -p 9001 -c .storybook/web",
     "storybook-native": "start-storybook -p 9002 -c .storybook/native",
     "test": "gulp test",
