+++ src/native/index.js
@@ -2,7 +2,7 @@
 /* eslint-disable react/require-extension */
 // Bootstrap environment
 const locales = require('./initialState').default.intl.locales;
-const polyfillLocales = require('../server/intl/polyfillLocales');
+const polyfillLocales = require('../common/intl/polyfillLocales');
 
 polyfillLocales(global, locales);
