--- src/server/.eslintrc
@@ -1,8 +1,0 @@
-{
-  "env": {
-    "node": true
-  },
-  "rules": {
-    "no-console": 0 // http://eslint.org/docs/rules/no-console#when-not-to-use-it
-  }
-}
